Refactor giant gui module
